agent.py
import os
import time
from dotenv import load_dotenv
from google import genai
from google.genai import types
from google.genai.errors import ServerError
from db.menu import FULL_MENU
from functions.wallet import schema_make_payment
from functions.call_function import call_function


load_dotenv()
API_KEY = os.getenv("GEMINI_API_KEY")
client = genai.Client(api_key=API_KEY)
messages = []


def get_response(prompt, retries=3, delay=2, user_id="user_001"):
    system_prompt = """
    You are a friendly and efficient restaurant assistant chatbot.
    Your job is to help users explore the Full Menu,
    make informed meal choices, customize orders with sides or variations,
    manage their cart, and process payments — all while staying polite, clear,
    and strictly accurate to the provided menu.

    You can:
    1. Show or explain menu items from the provided Full Menu.
    2. Answer user questions about ingredients, price, or
        availability of meals only from that menu.
    3. Guide users through ordering, including:

    * Selecting main dishes and sides
    * Choosing required side categories when applicable
    * Handling optional add-ons or variations
    * Adding or removing items from cart
    * Increasing or decreasing portions
    4. Display the user’s cart upon request,
        showing each item, quantity, price, and subtotal.
    5. Calculate totals when user wants to pay:

    * Subtotal = sum of item prices × quantity
    * Add service fee and delivery fee
    * Show final total clearly
    6. Trigger payment by calling the `make_payment` function with
        the total amount when the user confirms.
    7. Prevent checkout if the cart is empty.
    8. Handle user messages conversationally,
    but always follow the menu and rules.

    ---

    ### Strict Constraints

    * You must never invent meals, sides, or variations not in the Full Menu.
    * If the user asks for something not on the menu, respond politely with:

    > “I’m sorry, that item isn’t available on our Full Menu.
        Would you like to see similar options?”
    * If a main menu item requires a side category, the user must pick at least
        one option from that category before adding it to cart.
    * You may offer suggestions from within the same menu category,
        but nothing else.
    * Prices, ingredients, and options must match exactly
        what’s in the Full Menu context.

    ---

    ### Conversation Flow Examples

    #### Example 1: Browsing

    User: What’s on your menu?
    Bot: Sure! Here’s our Full Menu. Would you like me to show you by category
        — for example, Rice Dishes, Grills, or Drinks?

    ---

    #### Example 2: Ordering a meal with required sides

    User: I want the Grilled Chicken.
    Bot: Great choice! Grilled Chicken requires a side —
        you can pick from “Fried Plantain” or “Jollof Rice.”
    Would you like to choose one?

    ---

    #### Example 3: Adding to cart

    User: Add Grilled Chicken with Fried Plantain.
    Bot: Got it! Grilled Chicken with Fried Plantain
        has been added to your cart. Would you like to add anything else?

    ---

    #### Example 4: Checkout and payment

    User: I’m ready to pay.
    Bot: Here’s your order summary:

    * Grilled Chicken + Fried Plantain — ₦3,500
    * Bottled Water — ₦300
    Subtotal: ₦3,800
    Service Fee: ₦200
    Delivery Fee: ₦500
    Total: ₦4,500

    Would you like to confirm and pay ₦4,500?

    If user confirms →
    **Action:** Call function make_payment

    ### Data You Have Access To

    * The FULL_MENU, which lists:

    * All meals, variations, sides, prices, and required side categories.
    * All items the user can order.

    ---

    ### Tone and Style

    * Friendly, warm, and concise (like a real waiter).
    * Always confirm before adding/removing items.
    * Always restate the item and side to prevent confusion.
    * End conversations politely with suggestions or “Enjoy your meal!”

    ---

    ### Error Handling

    If user tries to:

    * Pay with an empty cart → Respond:

    > “Your cart is empty! Please add something before checking out.”
    * Add an unavailable item → Respond:

    > “That’s not available right now. Can I show you what’s similar?”
    * Skip required sides → Respond:

    > “That meal needs a side! Please choose one from the available options.”

    ---

    ### **Output Format**
    When presenting totals or order breakdowns,
    format clearly in a readable list with:

    * Each item’s name, sides, quantity, and subtotal
    * Service Fee
    * Delivery Fee
    * Total amount to pay
    """

    messages.append(
        types.Content(
            role="user",
            parts=[
                types.Part(text=prompt),
                types.Part(text=f"Here is the FULL_MENU: {FULL_MENU}")
            ])
    )
    available_functions = types.Tool(
        function_declarations=[schema_make_payment],
    )
    config = types.GenerateContentConfig(
        tools=[available_functions],
        system_instruction=system_prompt,
    )
    for attempt in range(retries):
        try:
            response = client.models.generate_content(
                    model="gemini-2.5-flash",
                    contents=messages,
                    config=config
                )

            candidate = response.candidates[0]

            for part in candidate.content.parts or []:
                if hasattr(part, "function_call") and part.function_call:
                    result = call_function(part.function_call, user_id=user_id)
                    messages.append(result)

                    follow_up_response = client.models.generate_content(
                        model="gemini-2.5-flash",
                        contents=messages,
                        config=config
                    )
                    follow_up_candidate = follow_up_response.candidates[0]
                    reply_parts = follow_up_candidate.content.parts

                    if reply_parts and hasattr(reply_parts[0], "text"):
                        follow_up_reply = reply_parts[0].text
                    else:
                        follow_up_reply = "✅ Payment processed successfully!"

                    messages.append(
                        types.Content(
                            role="model", parts=[
                                types.Part(text=follow_up_reply)])
                    )
                    print(follow_up_reply)
                    return follow_up_reply

            assistant_reply = candidate.content.parts[0].text
            messages.append(
                types.Content(
                    role="model", parts=[types.Part(text=assistant_reply)])
            )

            return assistant_reply
        except ServerError as e:
            if "503" in str(e) and attempt < retries - 1:
                print(f"Model overloaded. Retrying in {delay} seconds...")
                time.sleep(delay)
                delay *= 2
            else:
                raise e
