requirements.txt
google-genai==1.47.0
python-dotenv==1.2.1
streamlit==1.37.1